Use sealed classes for sync lines
