+++ packages/powersync_core/lib/src/bucket_storage.dart
@@ -347,105 +347,6 @@ class SyncDataBatch {
   SyncDataBatch(this.buckets);
 }
 
-class SyncBucketData {
-  final String bucket;
-  final List<OplogEntry> data;
-  final bool hasMore;
-  final String? after;
-  final String? nextAfter;
-
-  const SyncBucketData(
-      {required this.bucket,
-      required this.data,
-      this.hasMore = false,
-      this.after,
-      this.nextAfter});
-
-  SyncBucketData.fromJson(Map<String, dynamic> json)
-      : bucket = json['bucket'],
-        hasMore = json['has_more'] ?? false,
-        after = json['after'],
-        nextAfter = json['next_after'],
-        data =
-            (json['data'] as List).map((e) => OplogEntry.fromJson(e)).toList();
-
-  Map<String, dynamic> toJson() {
-    return {
-      'bucket': bucket,
-      'has_more': hasMore,
-      'after': after,
-      'next_after': nextAfter,
-      'data': data
-    };
-  }
-}
-
-class OplogEntry {
-  final String opId;
-
-  final OpType? op;
-
-  /// rowType + rowId uniquely identifies an entry in the local database.
-  final String? rowType;
-  final String? rowId;
-
-  /// Together with rowType and rowId, this uniquely identifies a source entry
-  /// per bucket in the oplog. There may be multiple source entries for a single
-  /// "rowType + rowId" combination.
-  final String? subkey;
-
-  final String? data;
-  final int checksum;
-
-  const OplogEntry(
-      {required this.opId,
-      required this.op,
-      this.subkey,
-      this.rowType,
-      this.rowId,
-      this.data,
-      required this.checksum});
-
-  OplogEntry.fromJson(Map<String, dynamic> json)
-      : opId = json['op_id'],
-        op = OpType.fromJson(json['op']),
-        rowType = json['object_type'],
-        rowId = json['object_id'],
-        checksum = json['checksum'],
-        data = json['data'] is String ? json['data'] : jsonEncode(json['data']),
-        subkey = json['subkey'] is String ? json['subkey'] : null;
-
-  Map<String, dynamic>? get parsedData {
-    return data == null ? null : jsonDecode(data!);
-  }
-
-  /// Key to uniquely represent a source entry in a bucket.
-  /// This is used to supersede old entries.
-  /// Relevant for put and remove ops.
-  String get key {
-    return "$rowType/$rowId/$subkey";
-  }
-
-  Map<String, dynamic> toJson() {
-    return {
-      'op_id': opId,
-      'op': op?.toJson(),
-      'object_type': rowType,
-      'object_id': rowId,
-      'checksum': checksum,
-      'subkey': subkey,
-      'data': data
-    };
-  }
-}
-
-class SqliteOp {
-  String sql;
-  List<dynamic> args;
-
-  SqliteOp(this.sql, this.args);
-}
-
 class SyncLocalDatabaseResult {
   final bool ready;
   final bool checkpointValid;
@@ -482,43 +383,4 @@ class ChecksumCache {
   Map<String, BucketChecksum> checksums;
 
   ChecksumCache(this.lastOpId, this.checksums);
-}
-
-enum OpType {
-  clear(1),
-  move(2),
-  put(3),
-  remove(4);
-
-  final int value;
-
-  const OpType(this.value);
-
-  static OpType? fromJson(String json) {
-    switch (json) {
-      case 'CLEAR':
-        return clear;
-      case 'MOVE':
-        return move;
-      case 'PUT':
-        return put;
-      case 'REMOVE':
-        return remove;
-      default:
-        return null;
-    }
-  }
-
-  String toJson() {
-    switch (this) {
-      case clear:
-        return 'CLEAR';
-      case move:
-        return 'MOVE';
-      case put:
-        return 'PUT';
-      case remove:
-        return 'REMOVE';
-    }
-  }
 }
