+++ packages/powersync_core/lib/src/streaming_sync.dart
@@ -53,7 +53,7 @@ class StreamingSyncImplementation implements StreamingSync {
 
   late final http.Client _client;
 
-  final StreamController<String?> _localPingController =
+  final StreamController<Null> _localPingController =
       StreamController.broadcast();
 
   final Duration retryDelay;
@@ -340,96 +340,19 @@ class StreamingSyncImplementation implements StreamingSync {
       }
 
       _updateStatus(connected: true, connecting: false);
-      if (line is Checkpoint) {
-        targetCheckpoint = line;
-        final Set<String> bucketsToDelete = {...bucketSet};
-        final Set<String> newBuckets = {};
-        for (final checksum in line.checksums) {
-          newBuckets.add(checksum.bucket);
-          bucketsToDelete.remove(checksum.bucket);
-        }
-        bucketSet = newBuckets;
-        await adapter.removeBuckets([...bucketsToDelete]);
-        _updateStatus(downloading: true);
-      } else if (line is StreamingSyncCheckpointComplete) {
-        final result = await adapter.syncLocalDatabase(targetCheckpoint!);
-        if (!result.checkpointValid) {
-          // This means checksums failed. Start again with a new checkpoint.
-          // TODO: better back-off
-          // await new Promise((resolve) => setTimeout(resolve, 50));
-          return false;
-        } else if (!result.ready) {
-          // Checksums valid, but need more data for a consistent checkpoint.
-          // Continue waiting.
-        } else {
-          appliedCheckpoint = targetCheckpoint;
-
-          _updateStatus(
-              downloading: false,
-              downloadError: _noError,
-              lastSyncedAt: DateTime.now());
-        }
-
-        validatedCheckpoint = targetCheckpoint;
-      } else if (line is StreamingSyncCheckpointDiff) {
-        // TODO: It may be faster to just keep track of the diff, instead of the entire checkpoint
-        if (targetCheckpoint == null) {
-          throw PowerSyncProtocolException(
-              'Checkpoint diff without previous checkpoint');
-        }
-        _updateStatus(downloading: true);
-        final diff = line;
-        final Map<String, BucketChecksum> newBuckets = {};
-        for (var checksum in targetCheckpoint.checksums) {
-          newBuckets[checksum.bucket] = checksum;
-        }
-        for (var checksum in diff.updatedBuckets) {
-          newBuckets[checksum.bucket] = checksum;
-        }
-        for (var bucket in diff.removedBuckets) {
-          newBuckets.remove(bucket);
-        }
-
-        final newCheckpoint = Checkpoint(
-            lastOpId: diff.lastOpId,
-            checksums: [...newBuckets.values],
-            writeCheckpoint: diff.writeCheckpoint);
-        targetCheckpoint = newCheckpoint;
-
-        bucketSet = Set.from(newBuckets.keys);
-        await adapter.removeBuckets(diff.removedBuckets);
-        adapter.setTargetCheckpoint(targetCheckpoint);
-      } else if (line is SyncBucketData) {
-        _updateStatus(downloading: true);
-        await adapter.saveSyncData(SyncDataBatch([line]));
-      } else if (line is StreamingSyncKeepalive) {
-        if (line.tokenExpiresIn == 0) {
-          // Token expired already - stop the connection immediately
-          invalidCredentialsCallback?.call().ignore();
-          break;
-        } else if (line.tokenExpiresIn <= 30) {
-          // Token expires soon - refresh it in the background
-          if (credentialsInvalidation == null &&
-              invalidCredentialsCallback != null) {
-            credentialsInvalidation = invalidCredentialsCallback!().then((_) {
-              // Token has been refreshed - we should restart the connection.
-              haveInvalidated = true;
-              // trigger next loop iteration ASAP, don't wait for another
-              // message from the server.
-              _localPingController.add(null);
-            }, onError: (_) {
-              // Token refresh failed - retry on next keepalive.
-              credentialsInvalidation = null;
-            });
+      switch (line) {
+        case Checkpoint():
+          targetCheckpoint = line;
+          final Set<String> bucketsToDelete = {...bucketSet};
+          final Set<String> newBuckets = {};
+          for (final checksum in line.checksums) {
+            newBuckets.add(checksum.bucket);
+            bucketsToDelete.remove(checksum.bucket);
           }
-        }
-      } else {
-        if (targetCheckpoint == appliedCheckpoint) {
-          _updateStatus(
-              downloading: false,
-              downloadError: _noError,
-              lastSyncedAt: DateTime.now());
-        } else if (validatedCheckpoint == targetCheckpoint) {
+          bucketSet = newBuckets;
+          await adapter.removeBuckets([...bucketsToDelete]);
+          _updateStatus(downloading: true);
+        case StreamingSyncCheckpointComplete():
           final result = await adapter.syncLocalDatabase(targetCheckpoint!);
           if (!result.checkpointValid) {
             // This means checksums failed. Start again with a new checkpoint.
@@ -447,7 +370,85 @@ class StreamingSyncImplementation implements StreamingSync {
                 downloadError: _noError,
                 lastSyncedAt: DateTime.now());
           }
-        }
+
+          validatedCheckpoint = targetCheckpoint;
+        case StreamingSyncCheckpointDiff():
+          // TODO: It may be faster to just keep track of the diff, instead of the entire checkpoint
+          if (targetCheckpoint == null) {
+            throw PowerSyncProtocolException(
+                'Checkpoint diff without previous checkpoint');
+          }
+          _updateStatus(downloading: true);
+          final diff = line;
+          final Map<String, BucketChecksum> newBuckets = {};
+          for (var checksum in targetCheckpoint.checksums) {
+            newBuckets[checksum.bucket] = checksum;
+          }
+          for (var checksum in diff.updatedBuckets) {
+            newBuckets[checksum.bucket] = checksum;
+          }
+          for (var bucket in diff.removedBuckets) {
+            newBuckets.remove(bucket);
+          }
+
+          final newCheckpoint = Checkpoint(
+              lastOpId: diff.lastOpId,
+              checksums: [...newBuckets.values],
+              writeCheckpoint: diff.writeCheckpoint);
+          targetCheckpoint = newCheckpoint;
+
+          bucketSet = Set.from(newBuckets.keys);
+          await adapter.removeBuckets(diff.removedBuckets);
+          adapter.setTargetCheckpoint(targetCheckpoint);
+        case SyncBucketData():
+          _updateStatus(downloading: true);
+          await adapter.saveSyncData(SyncDataBatch([line]));
+        case StreamingSyncKeepalive():
+          if (line.tokenExpiresIn == 0) {
+            // Token expired already - stop the connection immediately
+            invalidCredentialsCallback?.call().ignore();
+            break;
+          } else if (line.tokenExpiresIn <= 30) {
+            // Token expires soon - refresh it in the background
+            if (credentialsInvalidation == null &&
+                invalidCredentialsCallback != null) {
+              credentialsInvalidation = invalidCredentialsCallback!().then((_) {
+                // Token has been refreshed - we should restart the connection.
+                haveInvalidated = true;
+                // trigger next loop iteration ASAP, don't wait for another
+                // message from the server.
+                _localPingController.add(null);
+              }, onError: (_) {
+                // Token refresh failed - retry on next keepalive.
+                credentialsInvalidation = null;
+              });
+            }
+          }
+        case null: // Local ping
+          if (targetCheckpoint == appliedCheckpoint) {
+            _updateStatus(
+                downloading: false,
+                downloadError: _noError,
+                lastSyncedAt: DateTime.now());
+          } else if (validatedCheckpoint == targetCheckpoint) {
+            final result = await adapter.syncLocalDatabase(targetCheckpoint!);
+            if (!result.checkpointValid) {
+              // This means checksums failed. Start again with a new checkpoint.
+              // TODO: better back-off
+              // await new Promise((resolve) => setTimeout(resolve, 50));
+              return false;
+            } else if (!result.ready) {
+              // Checksums valid, but need more data for a consistent checkpoint.
+              // Continue waiting.
+            } else {
+              appliedCheckpoint = targetCheckpoint;
+
+              _updateStatus(
+                  downloading: false,
+                  downloadError: _noError,
+                  lastSyncedAt: DateTime.now());
+            }
+          }
       }
 
       if (haveInvalidated) {
@@ -458,7 +459,8 @@ class StreamingSyncImplementation implements StreamingSync {
     return true;
   }
 
-  Stream<Object?> streamingSyncRequest(StreamingSyncRequest data) async* {
+  Stream<StreamingSyncLine?> streamingSyncRequest(
+      StreamingSyncRequest data) async* {
     final credentials = await credentialsCallback();
     if (credentials == null) {
       throw CredentialsException('Not logged in');
@@ -498,7 +500,7 @@ class StreamingSyncImplementation implements StreamingSync {
       if (aborted) {
         break;
       }
-      yield parseStreamingSyncLine(line as Map<String, dynamic>);
+      yield StreamingSyncLine.fromJson(line as Map<String, dynamic>);
     }
   }
